+++ package/sdk/src/lib/config.ts
@@ -52,7 +52,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "1.0.0",
-  sdkVersion: "0.1.3",
-  genVersion: "2.436.3",
-  userAgent: "speakeasy-sdk/typescript 0.1.3 2.436.3 1.0.0 ryan-total-test-act",
+  sdkVersion: "0.1.4",
+  genVersion: "2.438.1",
+  userAgent: "speakeasy-sdk/typescript 0.1.4 2.438.1 1.0.0 ryan-total-test-act",
 } as const;